+++ app/src/main/AndroidManifest.xml
@@ -9,6 +9,23 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+
+        <activity android:name=".Activity.Activity_AddCourse"></activity>
+        <activity android:name=".Activity.Activity_TransactionHistory" />
+        <activity android:name=".Activity.Activity_MyWallet" />
+        <activity android:name=".Activity.Activity_AboutUs" />
+        <activity android:name=".Activity.Activity_HelpAndFeedback" />
+        <activity android:name=".Activity.Activity_InformationReminder" />
+        <activity android:name=".Activity.Activity_IdAuthentationSecond" />
+        <activity android:name=".Activity.Activity_IdAuthentationFirst" />
+        <activity android:name=".Activity.Activity_ChangeAccountInformation" />
+        <activity android:name=".Activity.Activity_AddCommodity" />
+        <activity android:name=".Activity.Activity_CommodityDetailInformation" />
+        <activity android:name=".Activity.Activity_AddNoteToCourse" />
+        <activity android:name=".Activity.Activity_CourseNote" />
+        <activity android:name=".Activity.Activity_CourseData" />
+        <activity android:name=".Activity.Activity_ClassGroupChat" />
+        <activity android:name=".Activity.Activity_ClassDetailInformation" />
         <activity
             android:name=".Activity.MainActivity"
             android:label="@string/app_name">
@@ -19,5 +36,7 @@
             </intent-filter>
         </activity>
     </application>
+
+
 
 </manifest>
