--- app/src/main/java/com/example/classchat/Activity/Activity_Main_Timetable.java
@@ -1,384 +1,0 @@
-package com.example.classchat.Activity;
-
-import android.content.DialogInterface;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.View.OnClickListener;
-import android.widget.ImageButton;
-import android.widget.LinearLayout;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.view.menu.MenuPopupHelper;
-import android.support.v7.widget.PopupMenu;
-
-import com.example.classchat.R;
-import com.example.classchat.model.MySubject;
-import com.example.classchat.model.SubjectRepertory;
-import com.example.library_activity_timetable.Activity_TimetableView;
-import com.example.library_activity_timetable.listener.ISchedule;
-import com.example.library_activity_timetable.listener.IWeekView;
-import com.example.library_activity_timetable.model.Schedule;
-import com.example.library_activity_timetable.view.WeekView;
-
-import java.lang.reflect.Field;
-import java.util.List;
-
-
-/**
- * 基础功能演示：
- * 1.周次选择栏
- * 2.透明背景
- * 3.点击监听
- * 4.颜色分配
- * 5.日期高亮
- * 6.日期计算
- * <p>
- * <p>
- * 该界面的代码注释会比较详细，建议从此处开始看起
- */
-public class Activity_Main_Timetable extends AppCompatActivity implements OnClickListener {
-
-    private static final String TAG = "Activity_Main_Timetable";
-
-    //控件
-    Activity_TimetableView mTimetableView;
-    WeekView mWeekView;
-
-    ImageButton moreButton;
-    LinearLayout layout;
-    TextView titleTextView;
-    List<MySubject> mySubjects;
-
-    //记录切换的周次，不一定是当前周
-    int target = -1;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity__main__timetable);
-
-        moreButton = findViewById(R.id.id_more);
-        moreButton.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                showPopmenu();
-            }
-        });
-
-        mySubjects = SubjectRepertory.loadDefaultSubjects2();
-        mySubjects.addAll(SubjectRepertory.loadDefaultSubjects());
-        titleTextView = findViewById(R.id.id_title);
-        layout = findViewById(R.id.id_layout);
-        layout.setOnClickListener(this);
-        initTimetableView();
-    }
-
-    /**
-     * 初始化课程控件
-     */
-    private void initTimetableView() {
-        //获取控件
-        mWeekView = findViewById(R.id.id_weekview);
-        mTimetableView = findViewById(R.id.id_timetableView);
-
-        //设置周次选择属性
-
-        mWeekView.source(mySubjects)
-                .curWeek(1)
-                .callback(new IWeekView.OnWeekItemClickedListener() {
-                    @Override
-                    public void onWeekClicked(int week) {
-                        int cur = mTimetableView.curWeek();
-                        //更新切换后的日期，从当前周cur->切换的周week
-                        mTimetableView.onDateBuildListener()
-                                .onUpdateDate(cur, week);
-                        mTimetableView.changeWeekOnly(week);
-                    }
-                })
-                .callback(new IWeekView.OnWeekLeftClickedListener() {
-                    @Override
-                    public void onWeekLeftClicked() {
-                        onWeekLeftLayoutClicked();
-                    }
-                })
-                .isShow(false)//设置隐藏，默认显示
-                .showView();
-
-        mTimetableView.source(mySubjects)
-                .curWeek(1)
-                .curTerm("大三下学期")
-                .maxSlideItem(10)
-                .monthWidthDp(30)
-                //透明度
-                //日期栏0.1f、侧边栏0.1f，周次选择栏0.6f
-                //透明度范围为0->1，0为全透明，1为不透明
-//                .alpha(0.1f, 0.1f, 0.6f)
-                .callback(new ISchedule.OnItemClickListener() {
-                    @Override
-                    public void onItemClick(View v, List<Schedule> scheduleList) {
-                        display(scheduleList);
-                    }
-                })
-                .callback(new ISchedule.OnItemLongClickListener() {
-                    @Override
-                    public void onLongClick(View v, int day, int start) {
-                        Toast.makeText(Activity_Main_Timetable.this,
-                                "长按:周" + day  + ",第" + start + "节",
-                                Toast.LENGTH_SHORT).show();
-                    }
-                })
-                .callback(new ISchedule.OnWeekChangedListener() {
-                    @Override
-                    public void onWeekChanged(int curWeek) {
-                        titleTextView.setText("第" + curWeek + "周");
-                    }
-                })
-                //旗标布局点击监听
-                .callback(new ISchedule.OnFlaglayoutClickListener() {
-                    @Override
-                    public void onFlaglayoutClick(int day, int start) {
-                        mTimetableView.hideFlaglayout();
-                        Toast.makeText(Activity_Main_Timetable.this,
-                                "点击了旗标:周" + (day + 1) + ",第" + start + "节",
-                                Toast.LENGTH_SHORT).show();
-                    }
-                })
-                .showView();
-        hideNonThisWeek();
-    }
-
-    /**
-     * 更新一下，防止因程序在后台时间过长（超过一天）而导致的日期或高亮不准确问题。
-     */
-    @Override
-    protected void onStart() {
-        super.onStart();
-        mTimetableView.onDateBuildListener()
-                .onHighLight();
-    }
-
-    /**
-     * 周次选择布局的左侧被点击时回调<br/>
-     * 对话框修改当前周次
-     */
-    protected void onWeekLeftLayoutClicked() {
-        final String items[] = new String[20];
-        int itemCount = mWeekView.itemCount();
-        for (int i = 0; i < itemCount; i++) {
-            items[i] = "第" + (i + 1) + "周";
-        }
-        target = -1;
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        builder.setTitle("设置当前周");
-        builder.setSingleChoiceItems(items, mTimetableView.curWeek() - 1,
-                new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
-                        target = i;
-                    }
-                });
-        builder.setPositiveButton("设置为当前周", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                if (target != -1) {
-                    mWeekView.curWeek(target + 1).updateView();
-                    mTimetableView.changeWeekForce(target + 1);
-                }
-            }
-        });
-        builder.setNegativeButton("取消", null);
-        builder.create().show();
-    }
-
-    /**
-     * 显示内容
-     *
-     * @param beans
-     */
-    protected void display(List<Schedule> beans) {
-        String str = "";
-        for (Schedule bean : beans) {
-            str += bean.getName() + ","+bean.getWeekList().toString()+","+bean.getStart()+","+bean.getStep()+"\n";
-        }
-        Toast.makeText(this, str, Toast.LENGTH_SHORT).show();
-    }
-
-    /**
-     * 显示弹出菜单
-     */
-
-    public void showPopmenu() {
-        PopupMenu popup = new PopupMenu(this, moreButton);
-        popup.getMenuInflater().inflate(R.menu.tabletime_menu, popup.getMenu());
-        popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-            public boolean onMenuItemClick(MenuItem item) {
-                switch (item.getItemId()) {
-                    case R.id.menu_add_course:
-                        addSubject();
-                        break;
-                    case R.id.menu_shownotthisweek:
-                        showNonThisWeek();
-                        break;
-
-                    case R.id.menu_hidenotthisweek:
-                        hideNonThisWeek();
-                        break;
-                    case R.id.menu_showweekend:
-                        showWeekends();
-                        break;
-                    case R.id.menu_hideweekend:
-                        hideWeekends();
-                        break;
-                    default:
-                        break;
-                }
-                return true;
-            }
-        });
-        try { //popupmenu显示icon的关键
-            Field mpopup=popup.getClass().getDeclaredField("mPopup");
-            mpopup.setAccessible(true);
-            MenuPopupHelper mPopup = (MenuPopupHelper) mpopup.get(popup);
-            mPopup.setForceShowIcon(true);
-        } catch (Exception e) {
-
-        }
-
-        popup.show();
-    }
-
-
-    @Override
-    public void onClick(View view) {
-
-        switch (view.getId()) {
-            case R.id.id_layout:
-                //如果周次选择已经显示了，那么将它隐藏，更新课程、日期
-                //否则，显示
-
-                if (mWeekView.isShowing()) {
-                    mWeekView.isShow(false);
-                    titleTextView.setTextColor(getResources().getColor(R.color.app_pinkpoint));
-                    int cur = mTimetableView.curWeek();
-                    mTimetableView.onDateBuildListener()
-                            .onUpdateDate(cur, cur);
-                    mTimetableView.changeWeekOnly(cur);
-                } else {
-                    mWeekView.isShow(true);
-                    titleTextView.setTextColor(getResources().getColor(R.color.app_gold));
-                }
-                break;
-        }
-    }
-
-    /**
-     * 删除课程
-     * 内部使用集合维护课程数据，操作集合的方法来操作它即可
-     * 最后更新一下视图（全局更新）
-     */
-    protected void deleteSubject() {
-        int size = mTimetableView.dataSource().size();
-        int pos = (int) (Math.random() * size);
-        if (size > 0) {
-            mTimetableView.dataSource().remove(pos);
-            mTimetableView.updateView();
-        }
-    }
-
-    /**
-     * 添加课程
-     * 内部使用集合维护课程数据，操作集合的方法来操作它即可
-     * 最后更新一下视图（全局更新）
-     */
-    protected void addSubject() {
-        Intent addsubjuct = new Intent(this, Activity_AddCourse.class);
-        startActivity(addsubjuct);
-        /*List<Schedule> dataSource = mTimetableView.dataSource();
-        int size = dataSource.size();
-        if (size > 0) {
-            Schedule schedule = dataSource.get(0);
-            dataSource.add(schedule);
-            mTimetableView.updateView();
-        }*/
-    }
-
-    /**
-     * 隐藏非本周课程
-     * 修改了内容的显示，所以必须更新全部（性能不高）
-     * 建议：在初始化时设置该属性
-     * <p>
-     * updateView()被调用后，会重新构建课程，课程会回到当前周
-     */
-    protected void hideNonThisWeek() {
-        mTimetableView.isShowNotCurWeek(false).updateView();
-    }
-
-    /**
-     * 显示非本周课程
-     * 修改了内容的显示，所以必须更新全部（性能不高）
-     * 建议：在初始化时设置该属性
-     */
-    protected void showNonThisWeek() {
-        mTimetableView.isShowNotCurWeek(true).updateView();
-    }
-
-    /**
-     * 设置侧边栏最大节次，只影响侧边栏的绘制，对课程内容无影响
-     *
-     * @param num
-     */
-    protected void setMaxItem(int num) {
-        mTimetableView.maxSlideItem(num).updateSlideView();
-    }
-
-
-    /**
-     * 显示WeekView
-     */
-
-    protected void showWeekView() {
-        mWeekView.isShow(true);
-    }
-
-    /**
-     * 隐藏WeekView
-     */
-
-    protected void hideWeekView() {
-        mWeekView.isShow(false);
-    }
-
-    /**
-     * 设置月份宽度
-     */
-    private void setMonthWidth() {
-        mTimetableView.monthWidthDp(50).updateView();
-    }
-
-    /**
-     * 设置月份宽度,默认40dp
-     */
-    private void resetMonthWidth() {
-        mTimetableView.monthWidthDp(40).updateView();
-    }
-
-    /**
-     * 隐藏周末
-     */
-    private void hideWeekends() {
-        mTimetableView.isShowWeekends(false).updateView();
-    }
-
-    /**
-     * 显示周末
-     */
-    private void showWeekends() {
-        mTimetableView.isShowWeekends(true).updateView();
-    }
-}
-
